more cleanup

+++ unix/dltest/pkgb.c
@@ -100,7 +100,7 @@ static int
 Pkgb_DemoObjCmd(
     ClientData dummy,		/* Not used. */
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,		/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     Tcl_Obj *first;
